costmap_3d: switch FCLFloat to double
From testing, using doubles is slightly faster for the distance
checks, although a bit slower for collision. Most clients will be
querying distance the most heavily (to optimize distances in path
processing), and the accuracy is needed when operating in a fixed
frame, as the querying may be relatively far from the origin.

+++ costmap_3d/include/costmap_3d/costmap_3d_query.h
@@ -172,7 +172,7 @@ private:
   // returns path to package file, or empty on error
   std::string getFileNameFromPackageURL(const std::string& url);
 
-  using FCLFloat = float;
+  using FCLFloat = double;
   using FCLRobotModel = fcl::BVHModel<fcl::OBBRSS<FCLFloat>>;
   using FCLRobotModelPtr = std::shared_ptr<FCLRobotModel>;
